Change int fields to SmallInteger

+++ cases/models.py
@@ -61,8 +61,8 @@ class Case(TimestampableModel):
     copy_of = models.ForeignKey("self", default=None, null=True, on_delete=models.DO_NOTHING)
     last_closed_at = models.DateTimeField(blank=True, null=True)
 
-    sla_days = models.IntegerField(null=False, blank=False, default=0)
-    sla_remaining_days = models.IntegerField(null=True, blank=True)
+    sla_days = models.PositiveSmallIntegerField(null=False, blank=False, default=0)
+    sla_remaining_days = models.SmallIntegerField(null=True, blank=True)
     sla_updated_at = models.DateTimeField(blank=True, null=True)
 
     objects = CaseManager()
